# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SolidFireGateway, Version=1.1.1.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/SolidFire/{cluster}/accounts/efficiency": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountID",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SolidFire/{cluster}/accounts/volumestats": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountID",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SolidFire/{cluster}/clusterstats/capacity": {
      "get": {
        "tags": [
          "ClusterStats"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SolidFire/{cluster}/clusterstats/performance": {
      "get": {
        "tags": [
          "ClusterStats"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/SolidFire/{cluster}/qospolicies": {
      "get": {
        "tags": [
          "Qos"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SolidFire/{cluster}/qospolicies/{policyID}": {
      "get": {
        "tags": [
          "Qos"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policyID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Qos"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policyID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyQoSPolicyParams"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyQoSPolicyParams"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyQoSPolicyParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Qos"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policyID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModifyQoSPolicyParams": {
        "type": "object",
        "properties": {
          "qosPolicyID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "qos": {
            "$ref": "#/components/schemas/QosUpdate"
          }
        },
        "additionalProperties": false
      },
...
      "QosUpdate": {
        "type": "object",
        "properties": {
          "minIOPS": {
            "type": "integer",
            "format": "int32"
          },
          "maxIOPS": {
            "type": "integer",
            "format": "int32"
          },
          "burstIOPS": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}